+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="com.example.pl_market.App"
+        android:name="com.example.plmarket.App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -15,19 +15,19 @@
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".MediatecaActivity"
+            android:name="com.example.plmarket.search.MediatecaActivity"
             android:exported="false" />
         <activity
-            android:name="search.SearchActivity"
+            android:name="com.example.plmarket.search.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".SettingsActivity"
+            android:name="com.example.plmarket.search.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".PlayerActivity"
+            android:name="com.example.plmarket.player.presentation.PlayerActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.plmarket.search.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
