--- app/src/main/java/com/example/pl_market/MediatecaActivity.kt
@@ -1,11 +1,0 @@
-package com.example.pl_market
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MediatecaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_mediateca)
-    }
-}
