--- app/src/main/java/com/example/pl_market/data/NetworkClient.kt
@@ -1,9 +1,0 @@
-package com.example.pl_market.data
-
-import com.example.pl_market.data.dto.Response
-
-interface NetworkClient {
-    var resultCode: Int
-
-    fun doRequest(dto: Any): Response
-}
