--- app/src/main/java/com/example/pl_market/data/dto/Response.kt
@@ -1,5 +1,0 @@
-package com.example.pl_market.data.dto
-
-open class Response() {
-    var resultCode = 0
-}
