--- app/src/main/java/com/example/pl_market/data/dto/TrackSearchRequest.kt
@@ -1,4 +1,0 @@
-package com.example.pl_market.data.dto
-
-data class TrackSearchRequest( val expression: String) {
-}
