--- app/src/main/java/com/example/pl_market/data/dto/TrackSearchResponse.kt
@@ -1,8 +1,0 @@
-package com.example.pl_market.data.dto
-
-class TrackSearchResponse(
-    val searchType: String,
-    val expression: String,
-    val results: List<TrackDto>
-) : Response() {
-}
