--- app/src/main/java/com/example/pl_market/domain/api/TrackRepository.kt
@@ -1,7 +1,0 @@
-package com.example.pl_market.domain.api
-
-import com.example.pl_market.domain.models.Track
-
-interface TrackRepository {
-    fun searchTrack(expression: String): List<Track>
-}
