--- app/src/main/java/com/example/pl_market/domain/models/Track.kt
@@ -1,16 +1,0 @@
-package com.example.pl_market.domain.models
-
-import android.os.Parcelable
-
-data class Track(
-    val trackName: String,
-    val artistName: String,
-    val trackTimeMillis: String,
-    val artworkUrl100: String?,
-    val trackId: Int,
-    val collectionName:String?,
-    val releaseDate: String?,
-    val primaryGenreName: String?,
-    val country: String?,
-    val previewUrl: String?
-) {}
