--- app/src/main/java/com/example/pl_market/domain/repository/PlayerRepository.kt
@@ -1,5 +1,0 @@
-package com.example.pl_market.domain.repository
-
-interface PlayerRepository {
-    fun startPlayer()
-}
