+++ app/src/main/java/com/example/plmarket/MediatecaActivity.kt
@@ -1,0 +1,12 @@
+package com.example.plmarket.search
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.pl_market.R
+
+class MediatecaActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_mediateca)
+    }
+}
