+++ app/src/main/java/com/example/plmarket/player/data/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.example.plmarket.player.data
+
+import com.example.plmarket.player.data.dto.Response
+
+interface NetworkClient {
+       fun doRequest(dto: Any): Response
+}
