+++ app/src/main/java/com/example/plmarket/player/data/dto/Response.kt
@@ -1,0 +1,3 @@
+package com.example.plmarket.player.data.dto
+
+open class Response(val resultCode: Int = 0)
