+++ app/src/main/java/com/example/plmarket/player/data/dto/TrackSearchRequest.kt
@@ -1,0 +1,3 @@
+package com.example.plmarket.player.data.dto
+
+data class TrackSearchRequest( val expression: String)
