+++ app/src/main/java/com/example/plmarket/player/data/dto/TrackSearchResponse.kt
@@ -1,0 +1,5 @@
+package com.example.plmarket.player.data.dto
+
+class TrackSearchResponse(
+    val results: List<TrackDto>
+) : Response()
