+++ app/src/main/java/com/example/plmarket/player/domain/api/TrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.plmarket.player.domain.api
+
+import com.example.plmarket.player.domain.models.Track
+
+interface TrackRepository {
+    fun searchTrack(expression: String): List<Track>
+}
