+++ app/src/main/java/com/example/plmarket/player/domain/models/Track.kt
@@ -1,0 +1,14 @@
+package com.example.plmarket.player.domain.models
+
+data class Track(
+    val trackName: String,
+    val artistName: String,
+    val trackTimeMillis: String,
+    val artworkUrl100: String?,
+    val trackId: Int,
+    val collectionName:String?,
+    val releaseDate: String?,
+    val primaryGenreName: String?,
+    val country: String?,
+    val previewUrl: String?
+)
